Client side rendering and editing of markdown syntax content stored in a tiddlyweb.
This makes it so a doubleclick on a html representation of a tiddler
will allow textarea editing. Once done the content is put to the server
and also rendered client side.

the html serialization is subclassed to add some behavior adding javascript

obviously there's little in the way of edit contention and the like
being handled here...