Yet another aggregation of code and plugins to experiment with packaging and making things in TiddlyWeb.
This is a "one to throw it away" of a thing called TiddlyHoster.

It is basically a system for shared hosting of tiddlywikis and other
tiddler content. If you are logged in you can go to a home page where
you can see your stuff and then access it. You can go to other
people's home pages and see their stuff (that they want you to see).
This home page is not done (and should probably come from an external
plugin).

If you are not logged in you are asked to either log in or register.
Registration is handled by the tiddlywebplugins.register plugin.

Output is managed by tiddlyweb.templates. This particular use skips
over the use of the HTMLPresenter by never setting 'tiddlyweb.title'
in the environment. It's becoming more and more clear that the
HTMLPresent is useful for the core urls, less so for ones added by
plugins.

What this stuff is designed to explore is:

* How well does the user model work/How can we use it best
* The pieces and parts needed to make a real TiddlyHoster
* Where we can find opportunities for abstraction.
* Can we store static content (like css) in tiddlers. What
  are the techniques for doing so.

At the moment what we're seeing is that the user database is
a bit light in terms of the information it is built to store
about a user. This was done pretty much on purpose in the
early design because proper user mgt requires a lot of complexity
if designed in. Now that we have a reached a stage where we want
some complexity in user handling, it appears that it will be possible
to manage that complexity in tiddlers and bags. That is: we can
use the tiddler store as a way for storing complex (and extensible)
information about a user. How this is to be accomplished remains to
be decided.

Storing css in tiddlers works fine. Other binary content has not
yet been explored (in this context).

Other things that will be explored:

* The user's "my stuff" page will presumably provide links to
  to tools for creating recipes, creating bags, adding content
  to bags (see tiddlywebplugins.wimporter).
* There should be a single click "host me a new wiki" interface.
  This can be facillitated by the "iframer" experiments, which
  host a proto-wiki in a tranformed iframe which can be commanded
  to upload when it has been give a new style or new set of plugins
  and the like.

Going a little more blue sky:

* In a perfect universe a user, while browsing other people's
  "stuff" should be able to say "I want to use this stuff in my stuff".
  Whether this builds a temporary recipe in a stepwise fashion or
  uses some kind of client side state that is eventually reified
  as a recipe is undef.
* Longer term a more complex fully cloud oriented TiddlyHoster
  will allow for the creation of distinct instances to which plugins
  can be installed. The tiddlywebplugins.plugins plugin (such a name!)
  is pursuing aspect of that functionality.

Issues:

* What happens when a user changes their name/openid/other identifier.
  Do all the policies have to be updated, or do we just say "don't
  do that"?
* Plugins which provide templating interfaces need to be flexible
  in the face of some other plugin wanting to override or horn in
  on their templates or flow. There's controller in some plugins
  that is desired without the view. This might mean that in addition
  to the template overriding that can be done with a
  plugin_local_templates dir, it might useful to be able to clobber
  plugin file names in the environment before get_template is reached.