[hoster] Manage some virtual role handling.
We replace the standard simple_cookie extractor with a sublcass
which takes its results and looks into the database to get two lists:

* all the people who this user follows
* all the people who follow this user

The current user then gets added to some, potentially many, roles:

 friend-of-<some user>
 <some user>-is-friend

based off the detailed generated above.

This provides a simple way of limiting access in meaningful but
social/fun way and demos how to do things with roles in the extractor.

It's also so FND can get at my python poetry porn.