First pass at an open auth client handling related to tiddlyweb.
This users twitter as the place we want to get stuff from.

For the moment the token information is stored in the user.note field,
but this is not good, will likely move it to a per-user tiddler
somewhere.

This should be reused. It's a scratchpad. It will get better.

I hope.